requirements.txt
mistralai
pdfkit
markdown
weasyprint
google-generativeai
#wkhtmltopdf